+++ ticketguardian/abstract/lazy_load_list.py
@@ -1,4 +1,3 @@
-from collections import defaultdict
 from math import floor
 
 from ticketguardian.constants import PAGE_SIZE
@@ -26,12 +25,12 @@ class LazyLoadMixin:
             self._data[i] = obj_list[i - start_ind]
 
     def _get_offset(self, ind):
-        page_number = floor(ind / (PAGE_SIZE))
+        page_number = int(floor(ind / (PAGE_SIZE)))
         return page_number * PAGE_SIZE
 
 
 class ResourceList(list, LazyLoadMixin):
-    def __init__(self, cls, size=None, data=None, slice_ind=0, *ext, **params):
+    def __init__(self, cls, size=None, slice_ind=0, *ext, **params):
         """
         cls: The class of the objects being listed
         size: The size of the ResourceList
@@ -46,13 +45,16 @@ class ResourceList(list, LazyLoadMixin):
         """
         self._cls = cls
         self._size = size or self._cls().get_resource_count(*ext, **params)
-        self._data = data or defaultdict(lambda: None)
+        self._data = params.pop('data', {})
         self._slice_ind = slice_ind
         self._ext = ext
         self._params = params
 
     def __repr__(self):
-        return F"<{self._cls.__name__} ResourceList: {self._size} objects>"
+        return "<{} ResourceList: {} objects>".format(
+            self._cls.__name__,
+            self._size
+        )
 
     def __iter__(self):
         return ResourceIterator(
@@ -77,7 +79,7 @@ class ResourceList(list, LazyLoadMixin):
                 data=self._data,
                 slice_ind=start,
                 *self._ext,
-                **self._params,
+                **self._params
             )
         if ind >= self._size:
             raise IndexError('list index out of range')
@@ -88,7 +90,7 @@ class ResourceList(list, LazyLoadMixin):
             # Add offsetting index
             ind += self._slice_ind
 
-        if self._data[ind] is None:
+        if self._data.get(ind) is None:
             offset = self._get_offset(ind)
             self._update_list(offset)
 
@@ -140,7 +142,7 @@ class ResourceIterator(LazyLoadMixin):
 
     def __next__(self):
         if self._ind < self._size:
-            if self._data[self._index] is None:
+            if self._data.get(self._index) is None:
                 offset = self._get_offset(self._index)
                 self._update_list(offset, self._lazy)
 
@@ -149,6 +151,10 @@ class ResourceIterator(LazyLoadMixin):
             return obj
         else:
             raise StopIteration
+
+    def next(self):
+        # Needed for python 2.7 compatibility
+        return self.__next__()
 
     @property
     def _index(self):