+++ ticketguardian/customer/customer.py
@@ -1,5 +1,13 @@
-from ticketguardian.abstract import RetrieveResourceMixin
+from ticketguardian.abstract import (
+    PostResourceMixin,
+    RetrieveResourceMixin,
+)
 
 
-class Customer(RetrieveResourceMixin):
+class Customer(
+    PostResourceMixin,
+    RetrieveResourceMixin):
     resource = 'customers'
+
+    def __init__(self):
+        super(Customer, self).__init__()