+++ ticketguardian/product/product.py
@@ -3,5 +3,10 @@ from ticketguardian.abstract import (
     RetrieveResourceMixin, )
 
 
-class Product(RetrieveResourceMixin, ListResourceMixin):
+class Product(
+    RetrieveResourceMixin,
+    ListResourceMixin):
     resource = "products"
+
+    def __init__(self):
+        super(Product, self).__init__()