+++ ticketguardian/quote/quote.py
@@ -1,11 +1,15 @@
-from ticketguardian.abstract import PostResourceMixin
+from ticketguardian.abstract import (
+    PatchResourceMixin,
+    PostResourceMixin,
+)
 
 
-class Quote(PostResourceMixin):
+class Quote(
+    PostResourceMixin):
     resource = "quote"
 
     def __init__(self, currency='USD', **params):
-        super().__init__()
+        super(Quote, self).__init__()
         self.create(
             instance=self,
             items=params.get('items'),